+++ src/ska_oso_services/odt/service/sbd_generator.py
@@ -1,4 +1,6 @@
 # pylint: disable=no-member
+from __future__ import annotations
+
 import logging
 from functools import singledispatch
 from random import randint